Big refactor of memory model into two arrays
First array is shared with NinjaTurtle data model, but has been bumped
to 16 floats (4x4 matrix) to enable easier use in the ES renderer. The
second array is for render-only data (colors) that NinjaTurtle doesn't
care about, so just an implementation details of Turgles.

Improved the buffer management in the process